convert constants scripts
