+++ src/constants/bucket.ts
@@ -1,2 +1,2 @@
 export const LEVANTE_BUCKET_URL = 'https://storage.googleapis.com/road-dashboard';
-export const LEVANTE_SURVEY_RESPONSES_KEY = 'levante-survey-responses';
+export const LEVANTE_SURVEY_RESPONSES_KEY = 'levante-survey-responses';
