+++ src/constants/consentTypes.ts
@@ -21,4 +21,6 @@ export const CONSENT_TYPES = Object.freeze({
   CONSENT_VIDEO_AUDIO_EYE_TRACKING_ES: 'consent-video-audio-eye-tracking-es',
   CONSENT_VIDEO_AUDIO_RECORDING: 'consent-video-audio-recording',
   CONSENT_VIDEO_AUDIO_RECORDING_ES: 'consent-video-audio-recording-es',
-});
+} as const);
+
+export type ConsentType = typeof CONSENT_TYPES[keyof typeof CONSENT_TYPES];
