+++ src/constants/csvExport.ts
@@ -3,4 +3,4 @@
  *
  * @constant {number} CSV_EXPORT_MAX_RECORD_COUNT - Maximum number of records allowed for CSV export.
  */
-export const CSV_EXPORT_MAX_RECORD_COUNT = 10000;
+export const CSV_EXPORT_MAX_RECORD_COUNT = 10000;
