+++ src/constants/mutationKeys.ts
@@ -6,4 +6,4 @@ export const TASK_ADD_MUTATION_KEY = 'task-add';
 export const TASK_UPDATE_MUTATION_KEY = 'task-update';
 export const TASK_VARIANT_ADD_MUTATION_KEY = 'task-variant-add';
 export const TASK_VARIANT_UPDATE_MUTATION_KEY = 'task-variant-update';
-export const USER_UPDATE_MUTATION_KEY = 'user-update';
+export const USER_UPDATE_MUTATION_KEY = 'user-update';
