--- src/constants/orgTypes.js
@@ -1,15 +1,0 @@
-export const ORG_TYPES = {
-  DISTRICTS: 'districts',
-  SCHOOLS: 'schools',
-  CLASSES: 'classes',
-  GROUPS: 'groups',
-  FAMILIES: 'families',
-};
-
-export const SINGULAR_ORG_TYPES = {
-  DISTRICTS: 'district',
-  SCHOOLS: 'school',
-  CLASSES: 'class',
-  GROUPS: 'group',
-  FAMILIES: 'family',
-};
