+++ src/constants/time.ts
@@ -5,4 +5,4 @@
  * @constant {number} oneSecondInMs - One second in milliseconds.
  */
 export const oneMinuteInMs = 60 * 1000;
-export const oneSecondInMs = 1000;
+export const oneSecondInMs = 1000;
