--- src/constants/toasts.js
@@ -1,7 +1,0 @@
-export const TOAST_SEVERITIES = Object.freeze({
-  SUCCESS: 'success',
-  INFO: 'info',
-  ERROR: 'error',
-});
-
-export const TOAST_DEFAULT_LIFE_DURATION = 3000;
